requirements.txt
Django==2.2.*
djangorestframework==3.10.*
markdown==3.1.*
django-filter==2.2.*
coreapi==2.3.*
djangorestframework_simplejwt==4.3.*
django-rest-swagger==2.2.*
django-cors-headers==3.2.*

Celery==4.3.*
redis==3.3.*

psycopg2==2.8.*
